// src/app/enums/status-dossier.ts
export enum StatusDossier {
  EN_COURS = 'EN_COURS',
  NOUVELLE_DEMANDE = 'NOUVELLE_DEMANDE',
  EN_ATTENTE_DE_COMPLEMENT = 'EN_ATTENTE_DE_COMPLEMENT',
  TRAITEE = 'TRAITEE',
  APPROUVE = 'APPROUVE',
  VALIDE_PAR_SGFG = 'VALIDE_PAR_SGFG',
  REJETE_PAR_SGFG = 'REJETE_PAR_SGFG',
  CLOTUREE = 'CLOTUREE',
  BROUILLON = 'BROUILLON',
  EN_ATTENTE_DE_QUALIFICATION = 'EN_ATTENTE_DE_QUALIFICATION',
  FRAUDE_VALIDEE = 'FRAUDE_VALIDEE',
  EN_ATTENTE_APPROBATION = 'EN_ATTENTE_APPROBATION',
  REJECTEE = "REJECTEE",
  INITIALE = "INITIALE",
  EXECUTEE = "EXECUTEE",
  VALIDER = "VALIDER",
  ANUULER = "ANUULER",
  EN_ATTENTE_DE_VALIDATION = "EN_ATTENTE_DE_VALIDATION",
  ACTIVER = "ACTIVER",
  DESACTIVER = "DESACTIVER"
}